Moved all trap handling functions to a TrapHandle static class
